Rehome PublicChannelSerializer and delete content/serializers.py
It serialises the /public/ channel-list response shape, not any content
viewset's, so it goes inline in its only consumer.
