--- kolibri/core/content/serializers.py
@@ -1,52 +1,0 @@
-from rest_framework import serializers
-
-from kolibri.core.content.models import ChannelMetadata
-from kolibri.core.fields import create_timezonestamp
-
-
-class PublicChannelSerializer(serializers.ModelSerializer):
-    included_languages = serializers.SerializerMethodField()
-    matching_tokens = serializers.SerializerMethodField("match_tokens")
-    language = serializers.SerializerMethodField()
-    icon_encoding = serializers.SerializerMethodField()
-    last_published = serializers.SerializerMethodField()
-
-    def get_language(self, instance):
-        if instance.root.lang is None:
-            return None
-
-        return instance.root.lang.lang_code
-
-    def get_icon_encoding(self, instance):
-        return instance.thumbnail
-
-    def get_included_languages(self, instance):
-        return list(instance.included_languages.all().values_list("id", flat=True))
-
-    def get_last_published(self, instance):
-        return (
-            None
-            if not instance.last_updated
-            else create_timezonestamp(instance.last_updated)
-        )
-
-    def match_tokens(self, channel):
-        return []
-
-    class Meta:
-        model = ChannelMetadata
-        fields = (
-            "id",
-            "name",
-            "language",
-            "included_languages",
-            "description",
-            "tagline",
-            "total_resource_count",
-            "version",
-            "published_size",
-            "last_published",
-            "icon_encoding",
-            "matching_tokens",
-            "public",
-        )
